move fetch-contentful as async action
